server/requirements.txt
azure-monitor-opentelemetry
pyyaml
contextvars
langchain-core
langchain
openai
langchain-openai
flask
flask-cors
asyncpg
azure-core
azure-storage-blob
azure-storage-queue
python_dotenv
pydantic
pytest
opentelemetry-sdk

